feat:tutorial imageslide->slide,slider login title 위치
Tutorial 사진 추가, 위치조정
imageslide를 slide, slider로 바꾸고 common으로
login 타이틀 위치 조정

Relatedto:#7, #6, #9
