+++ app/package.json
@@ -40,6 +40,7 @@
     "@react-native/typescript-config": "0.73.1",
     "@redux-devtools/extension": "^3.3.0",
     "@types/react": "^18.2.6",
+    "@types/react-native-vector-icons": "^6.4.18",
     "@types/react-test-renderer": "^18.0.0",
     "babel-jest": "^29.6.3",
     "eslint": "^8.19.0",
