+++ app/src/screens/TutorialScreen.js
@@ -1,5 +1,5 @@
 import React from "react";
-import ImageSlide from "../components/feat_yunsun/atom/ImageSlide";
+import ImageSlide from "../components/common/molecules/Slider";
 import Container from "../components/common/atom/Container";
 
 const TutorialScreen = ({ route, navigation, appState }) => {
